+++ packages/core/solidity/src/erc20.ts
@@ -5,7 +5,6 @@ import { addPauseFunctions } from './add-pausable';
 import { defineFunctions } from './utils/define-functions';
 import type { CommonOptions } from './common-options';
 import { withCommonDefaults, defaults as commonDefaults } from './common-options';
-import type { Upgradeable } from './set-upgradeable';
 import { setUpgradeable } from './set-upgradeable';
 import { setInfo } from './set-info';
 import { printContract } from './print';
@@ -90,7 +89,7 @@ export function buildERC20(opts: ERC20Options): ContractBuilder {
   addBase(c, allOpts.name, allOpts.symbol);
 
   if (allOpts.crossChainBridging) {
-    addCrossChainBridging(c, allOpts.crossChainBridging, allOpts.upgradeable, access);
+    addCrossChainBridging(c, allOpts.crossChainBridging, access);
   }
 
   if (allOpts.premint) {
@@ -305,25 +304,14 @@ function addFlashMint(c: ContractBuilder) {
   });
 }
 
-function addCrossChainBridging(
-  c: ContractBuilder,
-  crossChainBridging: 'custom' | 'superchain',
-  upgradeable: Upgradeable,
-  access: Access,
-) {
+function addCrossChainBridging(c: ContractBuilder, crossChainBridging: 'custom' | 'superchain', access: Access) {
   const ERC20Bridgeable = {
     name: 'ERC20Bridgeable',
-    path: `@openzeppelin/community-contracts/token/ERC20/extensions/ERC20Bridgeable.sol`,
+    path: `@openzeppelin/contracts/token/ERC20/extensions/draft-ERC20Bridgeable.sol`,
   };
 
   c.addParent(ERC20Bridgeable);
   c.addOverride(ERC20Bridgeable, supportsInterface);
-
-  if (upgradeable) {
-    throw new OptionsError({
-      crossChainBridging: 'Upgradeability is not currently supported with Cross-Chain Bridging',
-    });
-  }
 
   c.addOverride(ERC20Bridgeable, functions._checkTokenBridge);
   switch (crossChainBridging) {
@@ -345,13 +333,13 @@ function addCustomBridging(c: ContractBuilder, access: Access) {
   switch (access) {
     case false:
     case 'ownable': {
-      const addedBridgeImmutable = c.addVariable(`address public immutable TOKEN_BRIDGE;`);
+      const addedBridgeImmutable = c.addVariable(`address public tokenBridge;`);
       if (addedBridgeImmutable) {
-        c.addConstructorArgument({ type: 'address', name: 'tokenBridge' });
-        c.addConstructorCode(`require(tokenBridge != address(0), "Invalid TOKEN_BRIDGE address");`);
-        c.addConstructorCode(`TOKEN_BRIDGE = tokenBridge;`);
+        c.addConstructorArgument({ type: 'address', name: 'tokenBridge_' });
+        c.addConstructorCode(`require(tokenBridge_ != address(0), "Invalid tokenBridge_ address");`);
+        c.addConstructorCode(`tokenBridge = tokenBridge_;`);
       }
-      c.setFunctionBody([`if (caller != TOKEN_BRIDGE) revert Unauthorized();`], functions._checkTokenBridge, 'view');
+      c.setFunctionBody([`if (caller != tokenBridge) revert Unauthorized();`], functions._checkTokenBridge, 'view');
       break;
     }
     case 'roles': {